+++ app/controllers/ahoy/base_controller.rb
@@ -6,6 +6,7 @@ module Ahoy
     skip_around_action(*filters, raise: false)
 
     before_action :verify_request_size
+    before_action :check_params
     before_action :renew_cookies
 
     if respond_to?(:protect_from_forgery)
@@ -18,6 +19,13 @@ module Ahoy
       @ahoy ||= Ahoy::Tracker.new(controller: self, api: true)
     end
 
+    def check_params
+      if ahoy.send(:missing_params?)
+        logger.info "[ahoy] Missing required parameters"
+        render plain: "Missing required parameters\n", status: :bad_request
+      end
+    end
+
     # set proper ttl if cookie generated from JavaScript
     # approach is not perfect, as user must reload the page
     # for new cookie settings to take effect
@@ -28,7 +36,7 @@ module Ahoy
     def verify_request_size
       if request.content_length > Ahoy.max_content_length
         logger.info "[ahoy] Payload too large"
-        render plain: "Payload too large\n", status: 413
+        render plain: "Payload too large\n", status: :request_entity_too_large
       end
     end
   end